Share graph collection and tracing helpers
