+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/util.kt
@@ -254,27 +254,5 @@ internal fun <T> Sequence<T>.singleOrNullUnlessMultiple(
   return found
 }
 
-@JvmName("getAndAddSet")
-internal fun <K, V> MutableMap<K, MutableSet<V>>.getAndAdd(key: K, value: V) {
-  getOrInit(key).also { it.add(value) }
-}
-
-@JvmName("getAndAddList")
-internal fun <K, V> MutableMap<K, MutableList<V>>.getAndAdd(key: K, value: V) {
-  getOrInit(key).also { it.add(value) }
-}
-
-@IgnorableReturnValue
-@JvmName("getOrInitSet")
-internal fun <K, V> MutableMap<K, MutableSet<V>>.getOrInit(key: K): MutableSet<V> {
-  return getOrPut(key, ::mutableSetOf)
-}
-
-@IgnorableReturnValue
-@JvmName("getOrInitList")
-internal fun <K, V> MutableMap<K, MutableList<V>>.getOrInit(key: K): MutableList<V> {
-  return getOrPut(key, ::mutableListOf)
-}
-
 internal val ClassId.safePathString: String
   get() = asFqNameString().replace('.', '_')
