+++ metro-common/src/main/kotlin/dev/zacsweers/metro/compiler/MetroOptions.kt
@@ -2152,7 +2152,3 @@ private inline fun <reified T : Any> Any.expectAs(): T {
 private fun Any.diagnosticSeverity(): MetroOptions.DiagnosticSeverity {
   return MetroOptions.DiagnosticSeverity.valueOf(expectAs<String>().uppercase(Locale.US))
 }
-
-private fun <T, R> Sequence<T>.mapToSet(transform: (T) -> R): Set<R> {
-  return mapTo(mutableSetOf(), transform)
-}
